增加自定义校验条件
